+++ crypto/secp256k1/secp256k1.go
@@ -9,8 +9,9 @@ import (
 
 	"golang.org/x/crypto/ripemd160"
 
-	"github.com/tendermint/go-amino"
-	secp256k1wrap "github.com/tendermint/tendermint/crypto/secp256k1/wrapper"
+	secp256k1 "github.com/btcsuite/btcd/btcec"
+
+	amino "github.com/tendermint/go-amino"
 
 	"github.com/tendermint/tendermint/crypto"
 )
@@ -45,20 +46,10 @@ func (privKey PrivKeySecp256k1) Bytes() []byte {
 	return cdc.MustMarshalBinaryBare(privKey)
 }
 
-// Sign creates an ECDSA signature on curve Secp256k1, using SHA256 on the msg.
-func (privKey PrivKeySecp256k1) Sign(msg []byte) ([]byte, error) {
-	priv, _ := secp256k1wrap.PrivKeyFromBytes(secp256k1wrap.S256(), privKey[:])
-	sig, err := priv.Sign(crypto.Sha256(msg))
-	if err != nil {
-		return nil, err
-	}
-	return sig.Serialize(), nil
-}
-
 // PubKey performs the point-scalar multiplication from the privKey on the
 // generator point to get the pubkey.
 func (privKey PrivKeySecp256k1) PubKey() crypto.PubKey {
-	_, pubkeyObject := secp256k1wrap.PrivKeyFromBytes(secp256k1wrap.S256(), privKey[:])
+	_, pubkeyObject := secp256k1.PrivKeyFromBytes(secp256k1.S256(), privKey[:])
 	var pubkeyBytes PubKeySecp256k1
 	copy(pubkeyBytes[:], pubkeyObject.SerializeCompressed())
 	return pubkeyBytes
@@ -136,20 +127,6 @@ func (pubKey PubKeySecp256k1) Bytes() []byte {
 		panic(err)
 	}
 	return bz
-}
-
-func (pubKey PubKeySecp256k1) VerifyBytes(msg []byte, sig []byte) bool {
-	pub, err := secp256k1wrap.ParsePubKey(pubKey[:], secp256k1wrap.S256())
-	if err != nil {
-		return false
-	}
-	parsedSig, err := secp256k1wrap.ParseSignature(sig[:], secp256k1wrap.S256())
-	if err != nil {
-		return false
-	}
-	// Underlying library ensures that this signature is in canonical form, to
-	// prevent Secp256k1 malleability from altering the sign of the s term.
-	return parsedSig.Verify(crypto.Sha256(msg), pub)
 }
 
 func (pubKey PubKeySecp256k1) String() string {
