--- crypto/secp256k1/wrapper/doc.go
@@ -1,5 +1,0 @@
-// Package wrapper wraps github.com/btcsuite/btcd/btcec and slightly modifies
-// how signatures are serialized.
-//
-// The main difference between this package and the original btcec package is how ... TODO
-package wrapper
